Added example 3D Vis interaction papers
Responding to invitation by Liz to add example interaction technique papers from the 3D "SciVis" tradition.  There aren't as many as in other areas of VIS, but this is something we want to change and it may help to do that if we can highlight some good examples that do exist.  Interaction for 3D vis has been a focus of my group's research over the past 10 years, there are 3 best papers in the 4 suggested refs added to the list.  So, there is some recognition of quality, but I don't think the work is being picked up very often by the broader Vis community because SciVis is not known for interaction.  Tobias Isenberg has also published nice work on interaction for 3D Vis with his PhD student Lonni.  Tobias and I had an IEEE Computer review/position paper on interaction for SciVis a few years back https://ieeexplore.ieee.org/abstract/document/6515551 that has some good historical examples, like UNC's Nanomanipulator project, Bryson's virtual windtunnel, Hinkley's doll's head tangible UI, etc.  I'm not sure if those would be within the scope of your article, but it may be interesting to see some of the tradition of interaction for 3D vis that does exist.